refactor: change calculator to interface implementation
